requirements.txt
google-genai>=1.0.0
prefect>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
pytest>=7.0.0
pytest-cov>=4.0.0
PyYAML>=6.0
